requirements.txt
# aai python packages
#
# $ pip install -r requirements.txt

flask
gunicorn
starbug.coords
